+++ magick/distribute-cache.c
@@ -504,15 +504,15 @@ static MagickBooleanType OpenDistributeCache(SplayTreeInfo *registry,int file,
   */
   p=message;
   (void) memcpy(&image->storage_class,p,sizeof(image->storage_class));
-  p+=sizeof(image->storage_class);
+  p+=(ptrdiff_t) sizeof(image->storage_class);
   (void) memcpy(&image->colorspace,p,sizeof(image->colorspace));
-  p+=sizeof(image->colorspace);
+  p+=(ptrdiff_t) sizeof(image->colorspace);
   (void) memcpy(&image->channels,p,sizeof(image->channels));
-  p+=sizeof(image->channels);
+  p+=(ptrdiff_t) sizeof(image->channels);
   (void) memcpy(&image->columns,p,sizeof(image->columns));
-  p+=sizeof(image->columns);
+  p+=(ptrdiff_t) sizeof(image->columns);
   (void) memcpy(&image->rows,p,sizeof(image->rows));
-  p+=sizeof(image->rows);
+  p+=(ptrdiff_t) sizeof(image->rows);
   if (SyncImagePixelCache(image,exception) == MagickFalse)
     return(MagickFalse);
   status=AddValueToSplayTree(registry,(const void *) key,image);
@@ -560,15 +560,15 @@ static MagickBooleanType ReadDistributeCacheIndexes(SplayTreeInfo *registry,
     return(MagickFalse);
   q=message;
   (void) memcpy(&region.width,q,sizeof(region.width));
-  q+=sizeof(region.width);
+  q+=(ptrdiff_t) sizeof(region.width);
   (void) memcpy(&region.height,q,sizeof(region.height));
-  q+=sizeof(region.height);
+  q+=(ptrdiff_t) sizeof(region.height);
   (void) memcpy(&region.x,q,sizeof(region.x));
-  q+=sizeof(region.x);
+  q+=(ptrdiff_t) sizeof(region.x);
   (void) memcpy(&region.y,q,sizeof(region.y));
-  q+=sizeof(region.y);
+  q+=(ptrdiff_t) sizeof(region.y);
   (void) memcpy(&length,q,sizeof(length));
-  q+=sizeof(length);
+  q+=(ptrdiff_t) sizeof(length);
   p=GetVirtualPixels(image,region.x,region.y,region.width,region.height,
     exception);
   if (p == (const PixelPacket *) NULL)
@@ -618,15 +618,15 @@ static MagickBooleanType ReadDistributeCachePixels(SplayTreeInfo *registry,
     return(MagickFalse);
   q=message;
   (void) memcpy(&region.width,q,sizeof(region.width));
-  q+=sizeof(region.width);
+  q+=(ptrdiff_t) sizeof(region.width);
   (void) memcpy(&region.height,q,sizeof(region.height));
-  q+=sizeof(region.height);
+  q+=(ptrdiff_t) sizeof(region.height);
   (void) memcpy(&region.x,q,sizeof(region.x));
-  q+=sizeof(region.x);
+  q+=(ptrdiff_t) sizeof(region.x);
   (void) memcpy(&region.y,q,sizeof(region.y));
-  q+=sizeof(region.y);
+  q+=(ptrdiff_t) sizeof(region.y);
   (void) memcpy(&length,q,sizeof(length));
-  q+=sizeof(length);
+  q+=(ptrdiff_t) sizeof(length);
   p=GetVirtualPixels(image,region.x,region.y,region.width,region.height,
     exception);
   if (p == (const PixelPacket *) NULL)
@@ -682,15 +682,15 @@ static MagickBooleanType WriteDistributeCacheIndexes(SplayTreeInfo *registry,
     return(MagickFalse);
   p=message;
   (void) memcpy(&region.width,p,sizeof(region.width));
-  p+=sizeof(region.width);
+  p+=(ptrdiff_t) sizeof(region.width);
   (void) memcpy(&region.height,p,sizeof(region.height));
-  p+=sizeof(region.height);
+  p+=(ptrdiff_t) sizeof(region.height);
   (void) memcpy(&region.x,p,sizeof(region.x));
-  p+=sizeof(region.x);
+  p+=(ptrdiff_t) sizeof(region.x);
   (void) memcpy(&region.y,p,sizeof(region.y));
-  p+=sizeof(region.y);
+  p+=(ptrdiff_t) sizeof(region.y);
   (void) memcpy(&length,p,sizeof(length));
-  p+=sizeof(length);
+  p+=(ptrdiff_t) sizeof(length);
   q=GetAuthenticPixels(image,region.x,region.y,region.width,region.height,
     exception);
   if (q == (PixelPacket *) NULL)
@@ -740,15 +740,15 @@ static MagickBooleanType WriteDistributeCachePixels(SplayTreeInfo *registry,
     return(MagickFalse);
   p=message;
   (void) memcpy(&region.width,p,sizeof(region.width));
-  p+=sizeof(region.width);
+  p+=(ptrdiff_t) sizeof(region.width);
   (void) memcpy(&region.height,p,sizeof(region.height));
-  p+=sizeof(region.height);
+  p+=(ptrdiff_t) sizeof(region.height);
   (void) memcpy(&region.x,p,sizeof(region.x));
-  p+=sizeof(region.x);
+  p+=(ptrdiff_t) sizeof(region.x);
   (void) memcpy(&region.y,p,sizeof(region.y));
-  p+=sizeof(region.y);
+  p+=(ptrdiff_t) sizeof(region.y);
   (void) memcpy(&length,p,sizeof(length));
-  p+=sizeof(length);
+  p+=(ptrdiff_t) sizeof(length);
   q=GetAuthenticPixels(image,region.x,region.y,region.width,region.height,
     exception);
   if (q == (PixelPacket *) NULL)
@@ -1129,17 +1129,17 @@ MagickPrivate MagickBooleanType OpenDistributePixelCache(
     Serialize image attributes (see ValidatePixelCacheMorphology()).
   */
   (void) memcpy(p,&server_info->session_key,sizeof(server_info->session_key));
-  p+=sizeof(server_info->session_key);
+  p+=(ptrdiff_t) sizeof(server_info->session_key);
   (void) memcpy(p,&image->storage_class,sizeof(image->storage_class));
-  p+=sizeof(image->storage_class);
+  p+=(ptrdiff_t) sizeof(image->storage_class);
   (void) memcpy(p,&image->colorspace,sizeof(image->colorspace));
-  p+=sizeof(image->colorspace);
+  p+=(ptrdiff_t) sizeof(image->colorspace);
   (void) memcpy(p,&image->channels,sizeof(image->channels));
-  p+=sizeof(image->channels);
+  p+=(ptrdiff_t) sizeof(image->channels);
   (void) memcpy(p,&image->columns,sizeof(image->columns));
-  p+=sizeof(image->columns);
+  p+=(ptrdiff_t) sizeof(image->columns);
   (void) memcpy(p,&image->rows,sizeof(image->rows));
-  p+=sizeof(image->rows);
+  p+=(ptrdiff_t) sizeof(image->rows);
   count=dpc_send(server_info->file,p-message,message);
   if (count != (MagickOffsetType) (p-message))
     return(MagickFalse);
@@ -1206,17 +1206,17 @@ MagickPrivate MagickOffsetType ReadDistributePixelCacheIndexes(
   p=message;
   *p++='R';
   (void) memcpy(p,&server_info->session_key,sizeof(server_info->session_key));
-  p+=sizeof(server_info->session_key);
+  p+=(ptrdiff_t) sizeof(server_info->session_key);
   (void) memcpy(p,&region->width,sizeof(region->width));
-  p+=sizeof(region->width);
+  p+=(ptrdiff_t) sizeof(region->width);
   (void) memcpy(p,&region->height,sizeof(region->height));
-  p+=sizeof(region->height);
+  p+=(ptrdiff_t) sizeof(region->height);
   (void) memcpy(p,&region->x,sizeof(region->x));
-  p+=sizeof(region->x);
+  p+=(ptrdiff_t) sizeof(region->x);
   (void) memcpy(p,&region->y,sizeof(region->y));
-  p+=sizeof(region->y);
+  p+=(ptrdiff_t) sizeof(region->y);
   (void) memcpy(p,&length,sizeof(length));
-  p+=sizeof(length);
+  p+=(ptrdiff_t) sizeof(length);
   count=dpc_send(server_info->file,p-message,message);
   if (count != (MagickOffsetType) (p-message))
     return(-1);
@@ -1279,17 +1279,17 @@ MagickPrivate MagickOffsetType ReadDistributePixelCachePixels(
   p=message;
   *p++='r';
   (void) memcpy(p,&server_info->session_key,sizeof(server_info->session_key));
-  p+=sizeof(server_info->session_key);
+  p+=(ptrdiff_t) sizeof(server_info->session_key);
   (void) memcpy(p,&region->width,sizeof(region->width));
-  p+=sizeof(region->width);
+  p+=(ptrdiff_t) sizeof(region->width);
   (void) memcpy(p,&region->height,sizeof(region->height));
-  p+=sizeof(region->height);
+  p+=(ptrdiff_t) sizeof(region->height);
   (void) memcpy(p,&region->x,sizeof(region->x));
-  p+=sizeof(region->x);
+  p+=(ptrdiff_t) sizeof(region->x);
   (void) memcpy(p,&region->y,sizeof(region->y));
-  p+=sizeof(region->y);
+  p+=(ptrdiff_t) sizeof(region->y);
   (void) memcpy(p,&length,sizeof(length));
-  p+=sizeof(length);
+  p+=(ptrdiff_t) sizeof(length);
   count=dpc_send(server_info->file,p-message,message);
   if (count != (MagickOffsetType) (p-message))
     return(-1);
@@ -1341,7 +1341,7 @@ MagickPrivate MagickBooleanType RelinquishDistributePixelCache(
   p=message;
   *p++='d';
   (void) memcpy(p,&server_info->session_key,sizeof(server_info->session_key));
-  p+=sizeof(server_info->session_key);
+  p+=(ptrdiff_t) sizeof(server_info->session_key);
   count=dpc_send(server_info->file,p-message,message);
   if (count != (MagickOffsetType) (p-message))
     return(MagickFalse);
@@ -1408,17 +1408,17 @@ MagickPrivate MagickOffsetType WriteDistributePixelCacheIndexes(
   p=message;
   *p++='W';
   (void) memcpy(p,&server_info->session_key,sizeof(server_info->session_key));
-  p+=sizeof(server_info->session_key);
+  p+=(ptrdiff_t) sizeof(server_info->session_key);
   (void) memcpy(p,&region->width,sizeof(region->width));
-  p+=sizeof(region->width);
+  p+=(ptrdiff_t) sizeof(region->width);
   (void) memcpy(p,&region->height,sizeof(region->height));
-  p+=sizeof(region->height);
+  p+=(ptrdiff_t) sizeof(region->height);
   (void) memcpy(p,&region->x,sizeof(region->x));
-  p+=sizeof(region->x);
+  p+=(ptrdiff_t) sizeof(region->x);
   (void) memcpy(p,&region->y,sizeof(region->y));
-  p+=sizeof(region->y);
+  p+=(ptrdiff_t) sizeof(region->y);
   (void) memcpy(p,&length,sizeof(length));
-  p+=sizeof(length);
+  p+=(ptrdiff_t) sizeof(length);
   count=dpc_send(server_info->file,p-message,message);
   if (count != (MagickOffsetType) (p-message))
     return(-1);
@@ -1482,17 +1482,17 @@ MagickPrivate MagickOffsetType WriteDistributePixelCachePixels(
   p=message;
   *p++='w';
   (void) memcpy(p,&server_info->session_key,sizeof(server_info->session_key));
-  p+=sizeof(server_info->session_key);
+  p+=(ptrdiff_t) sizeof(server_info->session_key);
   (void) memcpy(p,&region->width,sizeof(region->width));
-  p+=sizeof(region->width);
+  p+=(ptrdiff_t) sizeof(region->width);
   (void) memcpy(p,&region->height,sizeof(region->height));
-  p+=sizeof(region->height);
+  p+=(ptrdiff_t) sizeof(region->height);
   (void) memcpy(p,&region->x,sizeof(region->x));
-  p+=sizeof(region->x);
+  p+=(ptrdiff_t) sizeof(region->x);
   (void) memcpy(p,&region->y,sizeof(region->y));
-  p+=sizeof(region->y);
+  p+=(ptrdiff_t) sizeof(region->y);
   (void) memcpy(p,&length,sizeof(length));
-  p+=sizeof(length);
+  p+=(ptrdiff_t) sizeof(length);
   count=dpc_send(server_info->file,p-message,message);
   if (count != (MagickOffsetType) (p-message))
     return(-1);
